+++ modules/analysis/aws-batch-demo/stack.py
@@ -160,14 +160,12 @@ class AwsBatchPipeline(Stack):
                     **{
                         "id": "AwsSolutions-IAM4",
                         "reason": "Managed Policies are for service account roles only",
-                        "applies_to": "*",
                     }
                 ),
                 NagPackSuppression(
                     **{
                         "id": "AwsSolutions-IAM5",
                         "reason": "Resource access restriced to ADDF resources",
-                        "applies_to": "*",
                     }
                 ),
             ],