+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - enforce HTTPS on OpenSearch
 - EFS FileSystemPolicies to improve security
 - removed all referneces to secrestsmanager in modulestack.yaml where not needed
+- core/eks module added region modifier to masterrole references
 
 ### **Removed**